# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "平行校园 Parallel Campus",
    "description": "知乎黑客松 2026 · 校园新锐季。公开数据 → 可编辑人格 → 分身在虚拟校园里替你先活一天。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/avatar/diary": {
      "get": {
        "tags": [
          "avatar"
        ],
        "summary": "Get Diary",
        "operationId": "get_diary_api_avatar_diary_get",
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/User"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiaryView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/avatar/whisper": {
      "post": {
        "tags": [
          "avatar"
        ],
        "summary": "Send Whisper",
        "description": "给分身留一句耳语（≤80 字）。下一 tick 强制决策（04 §5.4）。",
        "operationId": "send_whisper_api_avatar_whisper_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WhisperRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhisperResponseView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/avatar/report": {
      "get": {
        "tags": [
          "avatar"
        ],
        "summary": "Get Report",
        "operationId": "get_report_api_avatar_report_get",
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DiaryEntryView": {
        "properties": {
          "tick": {
            "type": "integer",
            "title": "Tick"
          },
          "time_label": {
            "type": "string",
            "title": "Time Label",
            "default": ""
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "tick",
          "type",
          "text"
        ],
        "title": "DiaryEntryView"
      },
      "DiaryView": {
        "properties": {
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/DiaryEntryView"
            },
            "type": "array",
            "title": "Entries"
          },
          "mood_series": {
            "items": {
              "$ref": "#/components/schemas/MoodPointView"
            },
            "type": "array",
            "title": "Mood Series"
          },
          "reflections": {
            "items": {
              "$ref": "#/components/schemas/ReflectionItemView"
            },
            "type": "array",
            "title": "Reflections"
          },
          "whispers": {
            "items": {
              "$ref": "#/components/schemas/WhisperItemView"
            },
            "type": "array",
            "title": "Whispers"
          }
        },
        "type": "object",
        "required": [
          "day"
        ],
        "title": "DiaryView"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoodPointView": {
        "properties": {
          "tick": {
            "type": "integer",
            "title": "Tick"
          },
          "valence": {
            "type": "number",
            "title": "Valence"
          },
          "arousal": {
            "type": "number",
            "title": "Arousal"
          }
        },
        "type": "object",
        "required": [
          "tick",
          "valence",
          "arousal"
        ],
        "title": "MoodPointView"
      },
...
      "ReflectionItemView": {
        "properties": {
          "tick": {
            "type": "integer",
            "title": "Tick"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "importance": {
            "type": "integer",
            "title": "Importance",
            "default": 8
          }
        },
        "type": "object",
        "required": [
          "tick",
          "text"
        ],
        "title": "ReflectionItemView"
      },
...
      "ReportFriendCharacter": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "avatar_key": {
            "type": "string",
            "title": "Avatar Key"
          },
          "kind": {
            "type": "string",
            "enum": [
              "player",
              "npc",
              "system"
            ],
            "title": "Kind"
          },
          "archetype": {
            "type": "string",
            "title": "Archetype",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "avatar_key",
          "kind"
        ],
        "title": "ReportFriendCharacter"
      },
      "ReportFriendView": {
        "properties": {
          "character": {
            "$ref": "#/components/schemas/ReportFriendCharacter"
          },
          "story": {
            "type": "string",
            "title": "Story"
          },
          "shared_topics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Shared Topics"
          },
          "affinity": {
            "type": "integer",
            "title": "Affinity"
          },
          "affinity_delta": {
            "type": "integer",
            "title": "Affinity Delta"
          },
          "is_human": {
            "type": "boolean",
            "title": "Is Human",
            "default": false
          },
          "zhihu_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zhihu Url"
          }
        },
        "type": "object",
        "required": [
          "character",
          "story",
          "affinity",
          "affinity_delta"
        ],
        "title": "ReportFriendView"
      },
      "ReportView": {
        "properties": {
          "day": {
            "type": "integer",
            "title": "Day"
          },
          "generated_tick": {
            "type": "integer",
            "title": "Generated Tick"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "top_friends": {
            "items": {
              "$ref": "#/components/schemas/ReportFriendView"
            },
            "type": "array",
            "title": "Top Friends"
          }
        },
        "type": "object",
        "required": [
          "day",
          "generated_tick",
          "summary"
        ],
        "title": "ReportView"
      },
...
      "User": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "display_name": {
            "type": "string",
            "maxLength": 24,
            "title": "Display Name"
          },
          "avatar_key": {
            "type": "string",
            "title": "Avatar Key",
            "default": "av_01"
          },
          "auth_kind": {
            "type": "string",
            "maxLength": 10,
            "title": "Auth Kind",
            "default": "dev"
          },
          "zhihu_uid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zhihu Uid"
          },
          "zhihu_url_token": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zhihu Url Token"
          },
          "zhihu_token_enc": {
            "anyOf": [
              {
                "type": "string",
                "contentMediaType": "application/octet-stream"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zhihu Token Enc"
          },
          "token_expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Token Expires At"
          },
          "is_judge": {
            "type": "boolean",
            "title": "Is Judge",
            "default": false
          },
          "judge_username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Judge Username"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "last_login_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Login At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "display_name"
        ],
        "title": "User"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WhisperItemView": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "tick": {
            "type": "integer",
            "title": "Tick"
          },
          "accepted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "text",
          "tick"
        ],
        "title": "WhisperItemView"
      },
      "WhisperRequest": {
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Text"
          }
        },
        "type": "object",
        "required": [
          "text"
        ],
        "title": "WhisperRequest"
      },
      "WhisperResponseView": {
        "properties": {
          "whisper_id": {
            "type": "string",
            "title": "Whisper Id"
          },
          "whispers_left_today": {
            "type": "integer",
            "title": "Whispers Left Today"
          }
        },
        "type": "object",
        "required": [
          "whisper_id",
          "whispers_left_today"
        ],
        "title": "WhisperResponseView"
      },
...
    }
  }
}